Use codec in subscriptions command.
