+++ snuba/subscriptions/worker.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import copy
 import itertools
-import json
 import time
 from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from datetime import timedelta
@@ -13,13 +12,11 @@ from snuba.reader import Result
 from snuba.request import Request
 from snuba.subscriptions.consumer import Tick
 from snuba.subscriptions.data import Subscription
-from snuba.utils.codecs import Encoder
 from snuba.utils.metrics.backends.abstract import MetricsBackend
 from snuba.utils.metrics.gauge import Gauge
 from snuba.utils.metrics.timer import Timer
 from snuba.utils.scheduler import ScheduledTask, Scheduler
 from snuba.utils.streams import Producer, Message, Topic
-from snuba.utils.streams.backends.kafka import KafkaPayload
 from snuba.utils.streams.batching import AbstractBatchWorker
 from snuba.web.query import parse_and_run_query
 
@@ -42,7 +39,7 @@ class SubscriptionWorker(
         dataset: Dataset,
         executor: ThreadPoolExecutor,
         schedulers: Mapping[int, Scheduler[Subscription]],
-        producer: Producer[KafkaPayload],
+        producer: Producer[SubscriptionTaskResult],
         topic: Topic,
         metrics: MetricsBackend,
         time_shift: Optional[timedelta] = None,
@@ -126,34 +123,6 @@ class SubscriptionWorker(
         # them to complete. Again, either the entire batch succeeds, or the
         # entire batch fails.
         for future in as_completed(
-            [
-                self.__producer.produce(
-                    self.__topic, subscription_task_result_encoder.encode(result)
-                )
-                for result in results
-            ]
+            [self.__producer.produce(self.__topic, result) for result in results]
         ):
             future.result()
-
-
-class SubscriptionTaskResultEncoder(Encoder[KafkaPayload, SubscriptionTaskResult]):
-    def encode(self, value: SubscriptionTaskResult) -> KafkaPayload:
-        subscription_id = str(value.task.task.identifier)
-        request, result = value.result
-        return KafkaPayload(
-            subscription_id.encode("utf-8"),
-            json.dumps(
-                {
-                    "version": 2,
-                    "payload": {
-                        "subscription_id": subscription_id,
-                        "request": {**request.body},
-                        "result": result,
-                        "timestamp": value.task.timestamp.isoformat(),
-                    },
-                }
-            ).encode("utf-8"),
-        )
-
-
-subscription_task_result_encoder = SubscriptionTaskResultEncoder()
